refactor: Move config files to corresponding directories
